[iOS] Fix incorrect default CRWWebViewScrollViewProxy properties
CRWWebViewScrollViewProxy initializes itself with a base UIScrollView
so any properties set on it before a real web view scroll view is set
are used (see the comment in CRWWebViewScrollViewProxy -init). Then,
when another scroll view is added to the proxy, all properties are
copied over from the stored scroll view to the newly added one.

However, there are a few properties for which a WKWebView.scrollView
(WKScrollView, a private class) and UIScrollView have different
defaults. This means that the property copying phase overwrites the
WKWebView.scrollView defaults with the UIScrollView defaults.

A better solution might be to track which properties are set, and only
copy over those properties in the copying phase, but that would be a
much more involved solution.

Bug: 1156224
Change-Id: Ie0ad8bad0439caf3fdb4ec37d03369c43d0a7100
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2580533
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836073}