Make NavigationSimulator to simulate sandbox/origin.
This patch makes two attributes of DidCommitNavigationParams:
- sandbox_flags
- origin
to be simulated more correctly.

Both arguments are meant to disapear soon and be replaced by the ones
computed from the browser process. In the meantime, we only check what
the browser and the renderer computed match, as close as possible.

This patch is required for:
https://chromium-review.googlesource.com/c/chromium/src/+/2579743/
Otherwise, the NavigationSimulator won't simulate them correctly and
there will be a mismatch.

Bug: 1041376
Change-Id: I87ee6ac583b1e8212a2e566934bf94f4e3b72a84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2584873
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836096}