[Android] Clear chosen data types from UI when sync is off
If MOBILE_IDENTITY_CONSISTENCY is not enabled sync data type states are
retained even if the user toggles 'Sync your Chrome data' off in
SyncAndServicesSettings page.
This leads to an UI error when the user is migrated while in this state
that shows that all data types are enabled to sync even though sync is
shown as turned off in ManageSyncSettings page.

This state is impossible to reach if MOBILE_IDENTITY_CONSISTENCY is enabled.
This cl creates a fix to migrate existing users who are in that state.

Recording of the faulty behavior can be found on the bug.

Bug: 1151853
Change-Id: I9716a5ac001da9048e7fa6e46f789fb3e5116709
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2581929
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836101}