Rework inheritance for SVG <pattern>
Switch SVGPatternElement to use an observer scheme similar to that used
by SVGGradientElement - i.e using an IdTargetObserver. Implement an
override of FindCycleFromSelf() on LayoutSVGResourcePattern to handle
the cycle-checking for the inheritance link.
LayoutSVGResourcePattern is refactored a little in the process to allow
reusing the attribute collection phase.
This drops the only remaining user of SVGResources::LinkedResource(), so
it and its associated code can be removed as well.

Bug: 1028063
Change-Id: I228f80fb707921ccab695ee8a573d17c0498fcbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2584307
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#836111}