[Autofill] Increase dropdown touch target
Before this patch, the touch target of the dropdowns in the card and
address editors were 24dp tall, which is shorter than the 48dp required
for accessibility.

This patch adds a 13dp padding on top and bottom of the dropdowns in the
card and address editor.

Because varying sizes of string values in the dropdown can wrap and
cause the dropdown height to increase, alignment with text fields on the
same row is difficult, so dropdowns are now prevented from sharing the
same line with text fields. This affects the US address format in
particular, where the "State" dropdown shared a line with the "ZIP Code"
textfield. Other forms that shared two dropdowns in the row (e.g.,
expiration month and year in card form) or two textfields in the row are
not affected.

After this patch, the touch target of the dropdowns in the card and
address editors is at least 48dp tall.

Bug: 977272
Change-Id: I9b99f514a7427af7ddf7f0963c724f2cdda0cb0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2586147
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836120}