Invalidate all backgrounds/scroll controls when we are invalidate subtree
We have a special background invalidation flag:
BackgroundNeedsFullPaintInvalidation(). When the flag is set, we'll
invalidate both the non-scrolling background and the scrolling
background.

However, when the flag is not set while we are invalidating the whole
subtree, previously we missed invalidating the scrolling background.

Scroll controls are similar.

Bug: 1155413
Change-Id: Id6c5ba4be127cce33d295379b9b444d7271385df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2585741
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836150}