Adds the kPdfRoot role and removes the kWebArea role
In the accessibility codebase, the kDocument role is used both for HTML elements
with an ARIA role of "document", as well as for the root node in Chromium's PDF viewer.

The above is confusing and leads to subtle bugs because:
1. Different native APIs might want to expose the root of a PDF document differently from the root of the web area,
e.g. Chrome's Automation API.
2. In fact, on most supported platforms, The PDF root is never exposed with the same role as the root of the web area.
For example on Linux ATK, the root of the web area has the ATK_ROLE_DOCUMENT_WEB whilst the root of the PDF document has the ATK_ROLE_DOCUMENT_FRAME.
3. The PDF plugin could be embedded in an iframe on any website.
PDFs are not always displayed as standalone documents.
4. This is future proof. Say in case we would need to add a predicate for navigating just to the top of the PDF,
instead of to the top of the whole webpage.

This patch fixes this by adding the new kPdfRoot role
In addition, the kWebArea role is no longer used and could be removed
simplifying all the places in the code where we had to check for either kRootWebArea or kWebArea roles.

Also took the opportunity to go through the whole codebase and ensure that the
kRootWebArea role was used appropriately.

AX-Relnotes: n/a.

R=dmazzoni@chromium.org, aleventhal@chromium.org

Change-Id: Ib8d1621b0003c51f9df48d19de357965d663cb1c
Bug: 1049261
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2556452
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Auto-Submit: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836163}