Fix comment to refer to correct function
The comment added in crrev.com/c/2407065 was copied from another line
without updating the function name, which makes it quite confusing.

I was looking at the code to see whether the technique from
crrev.com/c/2580440 would be applicable (it's not clear that it would
be an improvement in this case) but it did help me find this mistake.

Change-Id: If97941def14ec0d7d0d1ea191bf5e407243dddfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2586774
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836169}