Rename VisualViewport::MainFrame() to LocalMainFrame() for readability.
The function VisualViewport::MainFrame() is often used in if-statements
to test for the existence of a *local* mainframe. But since we could
also have a remote mainframe, or no mainframe at all, the name is
ambiguous. This CL renames the function to make it clear what condition
is actually being tested.

This CL is a cleanup-refactor prior to landing a fix for the associated
issue.

Bug: 1150977
Change-Id: Id1542e46516288b15306be6ba75ff1805206ddc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2585634
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836170}