Adjust ADB sideloading notification messages for FlexOrgs
Adjust ADB sideloading notification messages to support FlexOrgs by
adjusting the domain placeholder names and descriptions to make it
easier for the translators to understand the usage of the texts.

Additionally replace a call to
BrowserPolicyConnectorChromeOS::GetEnterpriseDisplayDomain with a call
to BrowserPolicyConnectorChromeOS::GetEnterpriseDomainManager.

Bug: b:172685929
Test: None
Change-Id: I1cdeff3e45730d2fed41997046734f3a98bf1ee3
Skip-Translation-Screenshots-Check: True
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2584775
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Brian Malcolm <bmalcolm@chromium.org>
Commit-Queue: Jana Grill <janagrill@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836193}