Implement a11y support for permission chip UI
Announce to screenreaders that a permission has requested if chip UI is
used. For prompt UI, the bubble is shown and it takes over the focus.
Chip UI is less intrusive so it must be explicitly annouced for screen
reader users. To avoid users not hearing the announcement, it would be
reannouced after 2 minutes.

Chip UI is now one of the panes (users can cycle through panes with F6). This way users can easily access it with keyboard which is very important for screenreader users.

Also to make sure that users don't miss pending permission request, added message about having pending permission request to tab announcement.

Bug: 1150836

Change-Id: I7118abcf0bed9702cf17d15a9bb59f9a1d8783b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2578843
Commit-Queue: Olesia Marukhno <olesiamarukhno@google.com>
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836198}