Tab Search: Fix margins and close button styling
This CL fixes it so the margins from the beginning of the tab search
item to the favicon match those from the end of the tab search item
to the close button.

This CL also hides the close button when the tab search item is not
hovered or focused to increase the space available to tab titles
and hostnames.

Before:
https://drive.google.com/file/d/1cmuKp4R_nVc7eAaljtIbFX2g0nwlIzKT/view?usp=sharing

After:
https://drive.google.com/file/d/1oa0lyapwxIGwtCpbxc61bJdPD03gfdVC/view?usp=sharing

Bug: 1128599
Change-Id: I6d49a6982483a87028f2d562884f752b82ea6cf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2582629
Reviewed-by: Roman Arora <romanarora@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836203}