Remove the invert bubble shown when Windows High Contrast is enabled
The ForcedColors feature is fully implemented and on track to ship in
M89. That brings Chrome to parity with Edge in how it handles Windows
High Contrast mode and it also gives web developers new CSS features
to detect and override it where needed.

InvertBubbleView was our previous solution - if a user had high
contrast mode enabled, we'd show this bubble to direct them to an
extension and theme to get a similar effect. Those recommendations
are now no longer necessary.

Bug: 970285
Change-Id: I0e0e78e489f2a40f1618055b7ebaea660e155e99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2586367
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836215}