[BreakoutBox] Invalidate frames when sending to the sink
This CL makes sure that the destroy method is automatically invoked on
VideoFrames when they are sent to a sink.
This automatic invalidation is important because there are cases when
frames are automatically sent to a sink without the possibility of
manual invalidation (e.g., when automatically piping a readable stream
to a writable stream).
This is also consistent with the automatic invalidation in
WebRTC Insertable Streams.

Drive-by: Re-enable a disabled test that flaked because of
a missing call to wait for a promise to be settled.

Bug: 1151913,1153092
Change-Id: I3bc40df3bb02a1ae9852b03af2c94898400a2d4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2587037
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836251}