Adding ViewsAXTreeManager output to chrome://accessibility
This change creates a ViewsAXTreeManager for each top-level Widget
behind the runtime flag "--enable-features=AccessibilityTreeForViews".
This change also displays the contents of the entire tree built by
ViewsAXTreeManager under chrome://accessibility under a new "Top
Level Windows" section that is only visible when the runtime flag
is enabled.

Much of the code changes here are converting methods that serialize
BrowserAccessibility nodes to instead take AXNode's. This was fairly
straightforward for most of these, however
AccessibilityTreeFormatterBlink::AddProperties had to be forked due to
all of the properties that depend on the delegate.

WidgetAXTreeIDMap is added to map between Widget objects and their
root AXTreeID.

In order to suppress build errors on ChromeOS and macOS, this change is
limited to non-ChromeOS Aura builds (Windows and Linux).

AX-Relnotes: Displays internal UI tree in chrome://accessibility.

Bug: 1049261
Change-Id: Id3918e177df360b284d186638a04d8be494508c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2402398
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836275}