Limit the use of KioskAppExternalLoader to Chrome App Kiosk only
KioskAppExternalLoader is used to load apps and extensions for a Chrome
App Kiosk session. It registers a callback and wait until app info is
updated.

This class should be only used in Chrome App Kiosk, but not in other two
Kiosk types. But according to the current implementation, the loader is
enabled unexpectedly and blocks the uninstallation of force-installed
extensions when they are removed by admins. Therefore, the check logic
should be limited to a proper scope.

Bug: 1153440
Change-Id: I8bf79fa553dfbea5e271b7f06fc6ff0dffb33495
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2575118
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Anatoliy Potapchuk <apotapchuk@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Commit-Queue: Anqing Zhao <anqing@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836300}