capture_mode: Keyboard shortcuts to create/resize a region.
Add the following keyboard shortcuts to a capture session.

Space: Creates a default region 1/12th the size of the root.
Tab: Cycles through and highlights the region and affordance circles.
  Also handles events as these were affecting the top window.
Arrow keys: Shifts the region if it is focused, resizes the region if
  a affordance circle is focused.

Not in this CL:
  - moving focus to capture bar/capture button, requires those widgets
    to be focusable
  - leave focus on mouse click and start tabbing from the last mouse
    click

Bug: 1154778
Test: manual, added test
Change-Id: Ib4adf61e93bc7ad759fb33bac3185d069de0da65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2575325
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836336}