[Autofill] Increase touch target for checkbox.
Before this patch, the touch target for the checkbox in the credit card
editor was 32dp, which is smaller than the 48dp required for
accessibility.

This patch sets the minimum height of 48dp for the checkbox in the
credit card editor.

After this patch, the touch target for the checkbox in the credit card
editor is exactly 48dp in height.

Bug: 977272
Change-Id: I25b5de2145fdadef6b2c411fcbbc4b781500e18a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2584462
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836338}