[CrOS PhoneHub] Do not show UI when host is pending
Before this CL, we would show the "connecting" UI when we were waiting
for a pending host to be set. This was misleading, since we were not
actually connecting at all. Even worse, we'd show this UI when the user
had accidentally picked the wrong device in the drop-down menu, perhaps
due to crbug/923594.

This CL updates Phone Hub so that it does not show any UI when we're in
this state. Once the phone is actually set up, then we show the
"connecting" UI.

Bug: 1157971, 1106937
Change-Id: I3cc896335b436849e05a8bc2d6b0adbbdc8edfb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2587563
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836382}