[CrOS PhoneHub] Silently add notifications after device suspension
This CL fixes an issue which re-shows notifications which have already
been shown to the user after the user has (1) suspended their device,
(2) locked their device, or (3) turned Bluetooth off.

When these situations occur, we silently add the notification to the
notification shade but do not pop it up. This fixes an issue which
caused notifications to appear spammy when we'd show them multiple
times.

Bug: 1157523, 1106937
Change-Id: I58efe0d96f226a75e07a137e22081637ebb524a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2587700
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836383}